+++ DynaLock/ISemaphore.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace DynaLock
+{
+    public interface ISemaphore : IDisposable
+    {
+        bool IsLockOwned();
+        void WaitOne(int millisecondsTimeout = 0);
+        void Release();
+    }
+}